+++ Modules/Microsoft365DSC/Modules/M365DSCPermissions.psm1
@@ -519,6 +519,9 @@ function Get-M365DSCApiServicePrincipal
 
     This application can then be used for Application Authentication.
 
+    With Type set to ManagedIdentity, the function assigns the permissions to an existing managed identity
+    instead. It creates no application, grants no consent and creates no credential.
+
     The provided permissions have to be as an array of hashtables, with Api set to the name of the API
     owning the permission and PermissionName set to the permission itself. The Api value is any API name a
     resource settings file uses, such as 'Office 365 Exchange Online' or 'Azure Service Management', one of
@@ -561,13 +564,15 @@ function Get-M365DSCApiServicePrincipal
     https://microsoft365dsc.com/user-guide/get-started/authentication-and-permissions/#power-apps-permissions
 
 .PARAMETER ApplicationName
-    Specifies the application display name.
+    Specifies the application display name. With Type set to ManagedIdentity, specifies the display name,
+    object id or client id of the managed identity.
 
 .PARAMETER Permissions
     Specifies permission definitions to assign.
 
 .PARAMETER Type
-    Specifies whether the app should use a secret or certificate credential.
+    Specifies whether the app should use a secret or certificate credential, or whether the permissions
+    are assigned to an existing managed identity.
 
 .PARAMETER MonthsValid
     Specifies the validity period in months for newly created credentials.
@@ -620,6 +625,10 @@ function Get-M365DSCApiServicePrincipal
     PS> $creds = Get-Credential
     PS> Update-M365DSCAzureAdApplication -ApplicationName 'Microsoft365DSC' -Permissions $((Get-M365DSCCompiledPermissionList -ResourceNameList (Get-M365DSCAllResources) -PermissionType Application -AccessType Read).Permissions) -Type Certificate -CreateSelfSignedCertificate -AdminConsent -MonthsValid 12 -Credential $creds -CertificatePath c:\Temp\M365DSC.cer
 
+.EXAMPLE
+    PS> $creds = Get-Credential
+    PS> Update-M365DSCAzureAdApplication -ApplicationName 'm365dsc-automation' -Permissions @(@{Api='Graph';PermissionName='Group.ReadWrite.All'},@{Api='Exchange';PermissionName='Exchange.ManageAsApp'}) -Type ManagedIdentity -Credential $creds
+
 .FUNCTIONALITY
     Public
 #>
@@ -640,7 +649,7 @@ function Update-M365DSCAzureAdApplication
 
         [Parameter(ParameterSetName = 'Secret')]
         [Parameter(ParameterSetName = 'Certificate')]
-        [ValidateSet('Secret', 'Certificate')]
+        [ValidateSet('Secret', 'Certificate', 'ManagedIdentity')]
         [System.String]
         $Type = 'Secret',
 
@@ -797,6 +806,100 @@ function Update-M365DSCAzureAdApplication
                 }
             }
         }
+        'ManagedIdentity'
+        {
+            Write-LogEntry -Message '  Assigning the permissions to a managed identity'
+            if ($AdminConsent)
+            {
+                Write-LogEntry -Message '  AdminConsent is ignored because permissions assigned to a managed identity need no consent.' -Type Warning
+            }
+        }
+    }
+
+    if ($Type -eq 'ManagedIdentity')
+    {
+        Write-LogEntry ' '
+        Write-LogEntry 'Checking existence of managed identity'
+        $identityFilter = "servicePrincipalType eq 'ManagedIdentity' and displayName eq '$($ApplicationName -replace "'", "''")'"
+        if ([System.Guid]::TryParse($ApplicationName, [ref][System.Guid]::Empty))
+        {
+            $identityFilter = "servicePrincipalType eq 'ManagedIdentity' and (id eq '$ApplicationName' or appId eq '$ApplicationName')"
+        }
+
+        $identities = @(Get-MgServicePrincipal -Filter $identityFilter -ErrorAction SilentlyContinue)
+        if ($identities.Count -eq 0)
+        {
+            Write-LogEntry -Message "No managed identity '$ApplicationName' found." -Type Error
+            return
+        }
+        if ($identities.Count -gt 1)
+        {
+            Write-LogEntry -Message "Multiple managed identities named '$ApplicationName' found. Specify the object id or client id of the managed identity instead." -Type Error
+            return
+        }
+
+        $identity = $identities[0]
+        Write-LogEntry "  Managed identity '$($identity.DisplayName)' found"
+
+        Write-LogEntry ' '
+        Write-LogEntry 'Checking managed identity permissions'
+        $assignmentsUri = "/v1.0/servicePrincipals/$($identity.Id)/appRoleAssignments"
+        $existingAssignments = Get-M365DSCRawGraphCollection -Uri $assignmentsUri
+        $servicePrincipalCache = @{}
+        foreach ($permission in $Permissions)
+        {
+            if ([System.String]::IsNullOrWhiteSpace($permission.Api) -or [System.String]::IsNullOrWhiteSpace($permission.PermissionName))
+            {
+                Write-LogEntry "Specified permission is invalid $(Convert-M365DscHashtableToString -Hashtable $permission)" -Type Warning
+                continue
+            }
+            Write-LogEntry "  Checking permission '$($permission.Api)\$($permission.PermissionName)'"
+
+            $svcprincipal = Get-M365DSCApiServicePrincipal -ApiName ($permission.Api) -Cache $servicePrincipalCache
+            if ($null -eq $svcprincipal)
+            {
+                Write-LogEntry "    No service principal found for API '$($permission.Api)'. Grant '$($permission.PermissionName)' manually." -Type Warning
+                continue
+            }
+
+            $appRoleId = ($svcprincipal.AppRoles | Where-Object -Property Value -EQ $permission.PermissionName).Id
+            if ($null -eq $appRoleId -and [System.Guid]::TryParse($permission.PermissionName, [ref][System.Guid]::Empty))
+            {
+                $appRoleId = $permission.PermissionName
+            }
+            if ($null -eq $appRoleId)
+            {
+                Write-LogEntry "    API '$($permission.Api)' has no application permission '$($permission.PermissionName)'." -Type Warning
+                continue
+            }
+
+            if ($null -ne ($existingAssignments | Where-Object -FilterScript { $_.resourceId -eq $svcprincipal.Id -and $_.appRoleId -eq $appRoleId }))
+            {
+                Write-LogEntry "    Permission '$($permission.Api)\$($permission.PermissionName)' already assigned to the managed identity!"
+                continue
+            }
+
+            try
+            {
+                $null = Invoke-MgGraphRequest -Method POST -Uri $assignmentsUri -Body @{
+                    principalId = $identity.Id
+                    resourceId  = $svcprincipal.Id
+                    appRoleId   = $appRoleId
+                } -ErrorAction Stop
+                Write-LogEntry "    Permission '$($permission.Api)\$($permission.PermissionName)' assigned to the managed identity"
+            }
+            catch
+            {
+                Write-LogEntry "    Error while assigning permission '$($permission.Api)\$($permission.PermissionName)': $($_.Exception.Message)" -Type Error
+            }
+        }
+
+        Write-LogEntry ' '
+        Write-LogEntry "Managed identity client id: $($identity.AppId)"
+        Write-LogEntry ' '
+        Write-LogEntry 'NOTE: Make sure you add the managed identity to the required Microsoft 365 (e.g. Global Admin) or Exchange (e.g. Organization Management) role groups as well!'
+        Write-LogEntry '      See the documentation for any required permissions.'
+        return
     }
 
     Write-LogEntry ' '